+++ Riot/Managers/Settings/Shared/RiotSharedSettings.swift
@@ -52,6 +52,34 @@ class RiotSharedSettings: NSObject {
 
     // MARK: - Public
 
+    // MARK: Integration provisioning
+
+    var hasIntegrationProvisioningEnabled: Bool {
+        return getIntegrationProvisioning()?.enabled ?? true
+    }
+
+    func getIntegrationProvisioning() -> RiotSettingIntegrationProvisioning? {
+        guard let integrationProvisioningDict = getAccountData(forEventType: Settings.integrationProvisioning) else {
+            return nil
+        }
+
+        return try? serializationService.deserialize(integrationProvisioningDict)
+    }
+
+    @discardableResult
+    func setIntegrationProvisioning(enabled: Bool,
+                                    success: @escaping () -> Void,
+                                    failure: @escaping (Error?) -> Void)
+        -> MXHTTPOperation? {
+
+        // Update only the "widgets" field in the account data
+        var integrationProvisioningDict = getAccountData(forEventType: Settings.integrationProvisioning) ?? [:]
+        integrationProvisioningDict[RiotSettingIntegrationProvisioning.CodingKeys.enabled.rawValue] = enabled
+
+        return session.setAccountData(integrationProvisioningDict, forType: Settings.integrationProvisioning, success: success, failure: failure)
+    }
+
+
     // MARK: Allowed widgets
     func permission(for widget: Widget) -> WidgetPermission {
         guard let allowedWidgets = getAllowedWidgets() else {
@@ -66,15 +94,11 @@ class RiotSharedSettings: NSObject {
             return nil
         }
 
-        do {
-            let allowedWidgets: RiotSettingAllowedWidgets = try serializationService.deserialize(allowedWidgetsDict)
-            return allowedWidgets
-        } catch {
-            return nil
-        }
+        return try? serializationService.deserialize(allowedWidgetsDict)
     }
 
-    @discardableResult func setPermission(_ permission: WidgetPermission,
+    @discardableResult
+    func setPermission(_ permission: WidgetPermission,
                                           for widget: Widget,
                                           success: @escaping () -> Void,
                                           failure: @escaping (Error?) -> Void)