+++ CHANGES.rst
@@ -8,6 +8,7 @@ Improvements:
  * RiotSharedSettings: New class to handle user settings shared accross Riot apps.
  * Widgets: Check user permission before opening a widget (TODO design: #2833).
  * Widgets: Add a contextual menu to refresh, open outside, remove and revoke the permission (#2834).
+ * Settings: Add an option for disabling use of the integration manager (#2843).
 
 Changes in 0.10.2 (2019-11-15)
 ===============================================